correctly hardens editable_future_graphs #11570
